refactor: extract Quicken-specific APIs into skill plugin
Move 5 Quicken-specific MCP tools (list_accounts, navigate_to_account,
read_register_state, set_register_filter, open_reconcile) out of the
core bridge/server layer into a new skills/quicken/ plugin.

Architecture changes:
- Add skills/ package with SkillPlugin protocol and auto-discovery
- Make _WIN32_CLASS_TO_ROLE and _WIN32_INTERACTIVE_CLASSES extensible
  via register_class_role() and register_interactive_class()
- Remove ~20 Quicken-specific class names from core dicts; they are
  now registered by the Quicken skill at load time
- Remove 5 abstract stubs from uia_bridge.py
- Remove 5 mock stubs from mock_bridge.py
- Remove 5 tool definitions from server.py
- Remove ~400 lines of Quicken-specific bridge code from win_bridge.py

Generic tools (capture_screenshot, send_win32_message, wait_for_element,
check_window_state, dismiss_modal_overlay) remain in core.

The skill auto-loads on startup; 251 tests still pass, 30 skipped.
