+++ server/win_bridge.py
@@ -134,9 +134,10 @@ def _fetch_typed_descendants(root, include_display: bool = False):
 # Win32-native fast element enumeration (bypasses UIA COM layer entirely)
 # ---------------------------------------------------------------------------
 
-# Maps Win32 class names (lowercase) to UIA-like role strings.
+# Built-in Win32 class name → UIA role mappings (standard controls only).
+# Application-specific mappings are registered at runtime via
+# ``register_class_role()`` from skill plugins.
 _WIN32_CLASS_TO_ROLE: dict[str, str] = {
-    # Standard Win32 controls
     "button": "button",
     "edit": "edit",
     "static": "text",
@@ -168,44 +169,34 @@ _WIN32_CLASS_TO_ROLE: dict[str, str] = {
     "syspager": "toolbar",
     "header": "header",
     "sysheader32": "header",
-    # Quicken-specific custom window classes (observed via EnumChildWindows)
-    "qc_button": "button",
-    "qwcombobox": "combobox",
-    "qwpanel": "pane",
-    "qwiconDisplay".lower(): "image",
-    "qw_bag_toolbar": "toolbar",
-    "qw_main_toolbar": "toolbar",
-    "qwmenubar": "menubar",
-    "qwlistbox": "list",
-    "qwedit": "edit",
-    # Quicken transaction register classes
-    "qredit": "edit",                          # register entry field (date, payee, amount)
-    "qwclass_transactionlist": "list",         # main transaction grid
-    "qwclass_txtoolbar": "toolbar",            # transaction toolbar (Save, More actions, Split)
-    "qwscrollbar": "scrollbar",               # Quicken custom scrollbar
-    "qwinchild": "pane",                       # generic Quicken child container
-    "qwnavbtntray": "toolbar",                 # account bar nav tray
-    "qwacctbarholder": "pane",                 # account bar holder
-    "qwnavigator": "pane",                     # left sidebar navigator
-    "qsidebar": "pane",                        # sidebar
-    "qwmdi": "pane",                           # MDI content pane
-    "mdifr": "pane",                           # MDI frame
 }
 
 # Win32 class names that have interactive actions (lowercase).
-_WIN32_INTERACTIVE_CLASSES: frozenset[str] = frozenset({
-    # Standard Win32
+# Extended at runtime by ``register_interactive_class()`` from skill plugins.
+_WIN32_INTERACTIVE_CLASSES: set[str] = {
     "button", "edit", "listbox", "combobox", "comboboxex32",
     "syslistview32", "listview", "systreeview32", "treeview",
     "systabcontrol32", "tabcontrol", "toolbarwindow32",
     "msctls_trackbar32", "msctls_updown32",
     "richedit20w", "richedit20a", "richedit50w", "rich edit",
     "sysmonthcal32", "sysdatetimepick32", "sysipaddress32",
-    # Quicken custom
-    "qc_button", "qwcombobox", "qwlistbox", "qwedit", "qwmenubar",
-    # Quicken transaction register
-    "qredit", "qwclass_transactionlist", "qwcombobox",
-})
+}
+
+
+def register_class_role(class_name: str, role: str) -> None:
+    """Register a custom Win32 class-name → role mapping.
+
+    Called by skill plugins to teach the bridge about app-specific controls.
+    """
+    _WIN32_CLASS_TO_ROLE[class_name.lower()] = role
+
+
+def register_interactive_class(class_name: str) -> None:
+    """Register a Win32 class-name as interactive.
+
+    Called by skill plugins to mark app-specific controls as keyboard-navigable.
+    """
+    _WIN32_INTERACTIVE_CLASSES.add(class_name.lower())
 
 # GWL_STYLE / WS_TABSTOP: controls with this style are keyboard-navigable.
 _WS_TABSTOP = 0x00010000
@@ -1606,453 +1597,6 @@ class WinUIABridge(UIABridge):
 
         return "", "none"
 
-    # ---------------------------------------------------------------------------
-    # Account navigation (Quicken-specific, Win32 combobox path)
-    # ---------------------------------------------------------------------------
-
-    _SKIP_ACCT_ITEMS = frozenset({"custom...", "qcombo_separator", ""})
-
-    def list_accounts(self) -> list[dict[str, Any]]:
-        """
-        Return all accounts visible in the 'All accounts' register combobox.
-
-        Strategy
-        --------
-        1. Find all ``qwcombobox`` controls in the current window.
-        2. Locate the one named "All accounts" (or the leftmost one that
-           contains non-date, non-type items).
-        3. Read its item list via ``CB_GETLBTEXT``.
-        4. Filter out separators, "Custom...", and empty strings.
-
-        Returns
-        -------
-        list of dict
-            Each entry has ``{"name": str, "combo_index": int}``.
-
-        Raises
-        ------
-        UIAError
-            If no account combobox can be located.
-        """
-        CB_GETCOUNT    = 0x0146
-        CB_GETLBTEXT   = 0x0148
-        CB_GETLBTEXTLEN= 0x0149
-        WM_GETTEXT     = 0x000D
-
-        import ctypes  # noqa: PLC0415
-
-        user32 = ctypes.windll.user32
-
-        from server.process_manager import get_process_manager  # noqa: PLC0415
-        pm = get_process_manager()
-        if not pm.attached:
-            raise TargetNotFoundError("Use select_window to attach to a target first.")
-
-        # Find all qwcombobox controls (platform-native: fast Win32 path)
-        all_els = self.find_all({
-            "roles": [], "has_actions": False, "named_only": False, "root": None,
-        })
-        combos = [e for e in all_els if e.get("class_name", "").lower() == "qwcombobox"]
-
-        # Prefer one explicitly named "All accounts"
-        def _hwnd_int(e: dict) -> int:
-            raw = e.get("hwnd", 0)
-            return int(raw, 16) if isinstance(raw, str) else int(raw)
-
-        def _get_items(h: int) -> list[str]:
-            count = user32.SendMessageW(h, CB_GETCOUNT, 0, 0)
-            out: list[str] = []
-            for i in range(min(count, 500)):
-                tlen = user32.SendMessageW(h, CB_GETLBTEXTLEN, i, 0)
-                if tlen <= 0:
-                    out.append("")
-                    continue
-                buf = ctypes.create_unicode_buffer(tlen + 1)
-                user32.SendMessageW(h, CB_GETLBTEXT, i, buf)
-                out.append(buf.value)
-            return out
-
-        acct_combo_h: int | None = None
-        for cb in combos:
-            nm = (cb.get("name") or "").lower()
-            if "account" in nm:
-                acct_combo_h = _hwnd_int(cb)
-                break
-
-        # Fallback: use leftmost combo that has non-date items
-        if acct_combo_h is None:
-            import ctypes.wintypes  # noqa: PLC0415
-            def _combo_x(e: dict) -> int:
-                r = ctypes.wintypes.RECT()
-                user32.GetWindowRect(_hwnd_int(e), ctypes.byref(r))
-                return r.left
-            for cb in sorted(combos, key=_combo_x):
-                items = _get_items(_hwnd_int(cb))
-                non_temporal = [
-                    it for it in items if it and not any(
-                        x in it.lower() for x in (
-                            "month", "year", "week", "day", "today",
-                            "quarter", "type", "income", "expense",
-                            "transfer", "all date",
-                        )
-                    )
-                ]
-                if len(non_temporal) > 1:
-                    acct_combo_h = _hwnd_int(cb)
-                    break
-
-        if acct_combo_h is None:
-            raise UIAError(
-                "No account combobox found. Navigate to a register view (e.g. SPENDING) first.",
-                code="ACCOUNT_COMBO_NOT_FOUND",
-            )
-
-        raw_items = _get_items(acct_combo_h)
-        result = []
-        for idx, name in enumerate(raw_items):
-            if name.lower() in self._SKIP_ACCT_ITEMS:
-                continue
-            result.append({"name": name, "combo_index": idx, "combo_hwnd": hex(acct_combo_h)})
-        return result
-
-    def navigate_to_account(self, account_name: str) -> dict[str, Any]:
-        """
-        Navigate the register view to a specific account.
-
-        Selects *account_name* in the "All accounts" toolbar combobox using
-        ``CB_SETCURSEL`` + ``WM_COMMAND(CBN_SELCHANGE)`` — no mouse interaction
-        required.
-
-        Parameters
-        ----------
-        account_name : str
-            Exact account name as returned by ``list_accounts()``.
-            Comparison is case-insensitive.
-
-        Returns
-        -------
-        dict
-            ``{"ok": True, "account": str, "combo_index": int}``
-
-        Raises
-        ------
-        UIAError
-            ``ACCOUNT_NOT_FOUND`` if the account name is not in the list.
-            ``ACCOUNT_COMBO_NOT_FOUND`` if no combobox can be located.
-        """
-        import ctypes  # noqa: PLC0415
-
-        CB_SETCURSEL  = 0x014E
-        CBN_SELCHANGE = 1
-        WM_COMMAND    = 0x0111
-
-        user32 = ctypes.windll.user32
-
-        accounts = self.list_accounts()
-        name_lower = account_name.lower()
-        match = next(
-            (a for a in accounts if a["name"].lower() == name_lower),
-            None,
-        )
-        if match is None:
-            available = [a["name"] for a in accounts]
-            raise UIAError(
-                f"Account {account_name!r} not found. Available: {available}",
-                code="ACCOUNT_NOT_FOUND",
-            )
-
-        combo_h = int(match["combo_hwnd"], 16)
-        idx = match["combo_index"]
-
-        user32.SendMessageW(combo_h, CB_SETCURSEL, idx, 0)
-        parent = user32.GetParent(combo_h)
-        ctrl_id = user32.GetDlgCtrlID(combo_h)
-        wparam = (CBN_SELCHANGE << 16) | (ctrl_id & 0xFFFF)
-        user32.SendMessageW(parent, WM_COMMAND, wparam, combo_h)
-
-        return {"ok": True, "account": match["name"], "combo_index": idx}
-
-    # ---------------------------------------------------------------------------
-    # Register state reading (reconcile / register view helpers)
-    # ---------------------------------------------------------------------------
-
-    def read_register_state(self) -> dict[str, Any]:
-        """
-        Return the current state of the visible transaction register.
-
-        Reads balance, transaction count, current account, and whether a
-        reconcile is in progress — all from Static/ComboBox controls, without
-        requiring access to owner-drawn rows.
-
-        Returns
-        -------
-        dict
-            ``{"ok": True, "account": str, "total": str, "count": str,
-               "reconcile_active": bool, "filter_text": str}``
-
-        Raises
-        ------
-        UIAError
-            ``REGISTER_NOT_FOUND`` if no TxList is visible.
-        """
-        import ctypes  # noqa: PLC0415
-        import ctypes.wintypes  # noqa: PLC0415
-
-        WM_GETTEXT     = 0x000D
-        WM_GETTEXTLENGTH = 0x000E
-        CB_GETCOUNT    = 0x0146
-        CB_GETCURSEL   = 0x0147
-        CB_GETLBTEXT   = 0x0148
-        CB_GETLBTEXTLEN= 0x0149
-
-        user32 = ctypes.windll.user32
-
-        from server.process_manager import get_process_manager  # noqa: PLC0415
-        pm = get_process_manager()
-        if not pm.attached:
-            raise TargetNotFoundError("Use select_window to attach first.")
-
-        root_hwnd = pm.attached.hwnd
-
-        def _read_text(h: int) -> str:
-            tlen = user32.SendMessageW(h, WM_GETTEXTLENGTH, 0, 0)
-            if tlen <= 0:
-                return ""
-            buf = ctypes.create_unicode_buffer(tlen + 1)
-            user32.SendMessageW(h, WM_GETTEXT, len(buf), buf)
-            return buf.value
-
-        def _combo_cur(h: int) -> str:
-            idx = user32.SendMessageW(h, CB_GETCURSEL, 0, 0)
-            if idx < 0:
-                return ""
-            tlen = user32.SendMessageW(h, CB_GETLBTEXTLEN, idx, 0)
-            if tlen <= 0:
-                return ""
-            buf = ctypes.create_unicode_buffer(tlen + 1)
-            user32.SendMessageW(h, CB_GETLBTEXT, idx, buf)
-            return buf.value
-
-        # Enumerate all children to find TxList, balance Static, count Static,
-        # account ComboBox, and filter Edit
-        all_ctrls: list[tuple[int, str, str]] = []
-        EnumCB = ctypes.WINFUNCTYPE(
-            ctypes.c_bool, ctypes.wintypes.HWND, ctypes.wintypes.LPARAM
-        )
-        def _cb(h: int, _: int) -> bool:
-            cls_buf = ctypes.create_unicode_buffer(256)
-            user32.GetClassNameW(h, cls_buf, 256)
-            t = _read_text(h)
-            all_ctrls.append((h, cls_buf.value.lower(), t))
-            return True
-        user32.EnumChildWindows(root_hwnd, EnumCB(_cb), 0)
-
-        txlist_h = next(
-            (h for h, c, _ in all_ctrls if c == "qwclass_transactionlist"
-             and user32.IsWindowVisible(h)), None
-        )
-        if txlist_h is None:
-            raise UIAError("No visible TxList found.", code="REGISTER_NOT_FOUND")
-
-        # Find the QWMDI ancestor of the TxList — balance statics are siblings
-        mdi_h = user32.GetParent(txlist_h)
-
-        mdi_children: list[tuple[int, str, str]] = []
-        def _cb2(h: int, _: int) -> bool:
-            cls_buf2 = ctypes.create_unicode_buffer(256)
-            user32.GetClassNameW(h, cls_buf2, 256)
-            t = _read_text(h)
-            mdi_children.append((h, cls_buf2.value.lower(), t))
-            return True
-        user32.EnumChildWindows(mdi_h, EnumCB(_cb2), 0)
-
-        # Total/balance — Static with numeric content (ignore "Total:" label)
-        def _looks_numeric(s: str) -> bool:
-            s2 = s.replace(",", "").replace(".", "").replace("-", "").strip()
-            return bool(s2) and s2.isdigit()
-
-        balance_static = next(
-            (t for h, c, t in mdi_children
-             if c == "static" and _looks_numeric(t) and user32.IsWindowVisible(h)),
-            "",
-        )
-
-        # Transaction count — Static like "N Transaction(s)"
-        count_static = next(
-            (t for h, c, t in mdi_children
-             if c == "static" and "transaction" in t.lower()
-             and user32.IsWindowVisible(h)),
-            "",
-        )
-
-        # Account combobox current selection (QWComboBox visible)
-        acct_combo_h = next(
-            (h for h, c, t in mdi_children
-             if c == "qwcombobox" and user32.IsWindowVisible(h)), None
-        )
-        current_account = _combo_cur(acct_combo_h) if acct_combo_h else ""
-
-        # Reconcile active? "C" button and "Reset" button both visible in header
-        c_btn_visible = any(
-            h for h, c, t in mdi_children
-            if c == "qc_button" and t == "C" and user32.IsWindowVisible(h)
-        )
-        reset_visible = any(
-            h for h, c, t in mdi_children
-            if c == "qc_button" and t == "Reset" and user32.IsWindowVisible(h)
-        )
-        reconcile_active = c_btn_visible and reset_visible
-
-        # Filter text (Edit control in reconcile header, y < TxList.top)
-        txlist_rect = ctypes.wintypes.RECT()
-        user32.GetWindowRect(txlist_h, ctypes.byref(txlist_rect))
-        filter_text = ""
-        for h, c, t in mdi_children:
-            if c != "edit" or not user32.IsWindowVisible(h):
-                continue
-            r = ctypes.wintypes.RECT()
-            user32.GetWindowRect(h, ctypes.byref(r))
-            if r.bottom <= txlist_rect.top:
-                filter_text = _read_text(h)
-                break
-
-        return {
-            "ok": True,
-            "account": current_account,
-            "total": balance_static,
-            "count": count_static,
-            "reconcile_active": reconcile_active,
-            "filter_text": filter_text,
-        }
-
-    def set_register_filter(self, text: str) -> dict[str, Any]:
-        """
-        Type *text* into the register search/filter box and return the
-        resulting transaction count.
-
-        Works in both normal register view (the global search bar) and the
-        reconcile register view (the inline filter Edit).
-
-        Parameters
-        ----------
-        text : str
-            Search term — e.g. ``"AMAZON"`` or ``"4/12/2026"``.
-            Pass ``""`` to clear the filter.
-
-        Returns
-        -------
-        dict
-            ``{"ok": True, "filter": str, "count": str}``
-
-        Raises
-        ------
-        UIAError
-            ``REGISTER_NOT_FOUND`` if no TxList is visible.
-            ``FILTER_EDIT_NOT_FOUND`` if no filter edit is found.
-        """
-        import ctypes  # noqa: PLC0415
-        import ctypes.wintypes  # noqa: PLC0415
-        import time  # noqa: PLC0415
-
-        WM_SETTEXT   = 0x000C
-        WM_GETTEXT   = 0x000D
-        WM_GETTEXTLENGTH = 0x000E
-        WM_CHAR      = 0x0102
-
-        user32 = ctypes.windll.user32
-
-        from server.process_manager import get_process_manager  # noqa: PLC0415
-        pm = get_process_manager()
-        if not pm.attached:
-            raise TargetNotFoundError("Use select_window to attach first.")
-
-        root_hwnd = pm.attached.hwnd
-
-        def _read_text(h: int) -> str:
-            tlen = user32.SendMessageW(h, WM_GETTEXTLENGTH, 0, 0)
-            if tlen <= 0:
-                return ""
-            buf = ctypes.create_unicode_buffer(tlen + 1)
-            user32.SendMessageW(h, WM_GETTEXT, len(buf), buf)
-            return buf.value
-
-        # Find TxList and its parent QWMDI
-        all_ctrls: list[tuple[int, str, str]] = []
-        EnumCB = ctypes.WINFUNCTYPE(
-            ctypes.c_bool, ctypes.wintypes.HWND, ctypes.wintypes.LPARAM
-        )
-        def _cb(h: int, _: int) -> bool:
-            cls_buf = ctypes.create_unicode_buffer(256)
-            user32.GetClassNameW(h, cls_buf, 256)
-            all_ctrls.append((h, cls_buf.value.lower(), _read_text(h)))
-            return True
-        user32.EnumChildWindows(root_hwnd, EnumCB(_cb), 0)
-
-        txlist_h = next(
-            (h for h, c, _ in all_ctrls if c == "qwclass_transactionlist"
-             and user32.IsWindowVisible(h)), None
-        )
-        if txlist_h is None:
-            raise UIAError("No visible TxList found.", code="REGISTER_NOT_FOUND")
-
-        txlist_rect = ctypes.wintypes.RECT()
-        user32.GetWindowRect(txlist_h, ctypes.byref(txlist_rect))
-
-        mdi_h = user32.GetParent(txlist_h)
-        mdi_children: list[tuple[int, str, str]] = []
-        def _cb2(h: int, _: int) -> bool:
-            cls_buf2 = ctypes.create_unicode_buffer(256)
-            user32.GetClassNameW(h, cls_buf2, 256)
-            mdi_children.append((h, cls_buf2.value.lower(), _read_text(h)))
-            return True
-        user32.EnumChildWindows(mdi_h, EnumCB(_cb2), 0)
-
-        # Find filter Edit above TxList
-        filter_h: int | None = None
-        for h, c, _ in mdi_children:
-            if c != "edit" or not user32.IsWindowVisible(h):
-                continue
-            r = ctypes.wintypes.RECT()
-            user32.GetWindowRect(h, ctypes.byref(r))
-            if r.bottom <= txlist_rect.top:
-                filter_h = h
-                break
-
-        if filter_h is None:
-            raise UIAError(
-                "No filter Edit found above TxList.",
-                code="FILTER_EDIT_NOT_FOUND",
-            )
-
-        # Write text via WM_SETTEXT and trigger change notification
-        buf = ctypes.create_unicode_buffer(text)
-        user32.SendMessageW(filter_h, WM_SETTEXT, 0, buf)
-        # Post EN_CHANGE to parent so Quicken re-filters
-        WM_COMMAND = 0x0111
-        EN_CHANGE = 0x0300
-        ctrl_id = user32.GetDlgCtrlID(filter_h)
-        parent = user32.GetParent(filter_h)
-        user32.PostMessageW(parent, WM_COMMAND, (EN_CHANGE << 16) | ctrl_id, filter_h)
-        time.sleep(0.4)  # allow Quicken to re-render
-
-        # Read updated count
-        count_static = next(
-            (t for h, c, t in mdi_children
-             if c == "static" and "transaction" in t.lower()
-             and user32.IsWindowVisible(h)),
-            "",
-        )
-        # Refresh count (values change after filter)
-        if count_static == "" and mdi_h:
-            for h, c, _ in mdi_children:
-                if c == "static" and user32.IsWindowVisible(h):
-                    t2 = _read_text(h)
-                    if "transaction" in t2.lower():
-                        count_static = t2
-                        break
-
-        return {"ok": True, "filter": text, "count": count_static}
-
     def capture_screenshot(
         self,
         hwnd: int | None = None,
@@ -2165,269 +1709,4 @@ class WinUIABridge(UIABridge):
             "width": width,
             "height": height,
             "format": "PNG",
-        }
-
-    def open_reconcile(
-        self,
-        account_name: str,
-        statement_date: str,
-        ending_balance: str,
-        service_charge: str = "",
-        service_date: str = "",
-        interest_earned: str = "",
-        interest_date: str = "",
-        timeout_ms: int = 5000,
-    ) -> dict[str, Any]:
-        """Open the Quicken reconcile dialog and enter statement details.
-
-        Sends WM_COMMAND 7203 to QFRAME to open the "Choose Reconcile Account"
-        dialog, selects *account_name*, clicks OK, fills in the statement date
-        and ending balance in the "Reconcile Details" dialog, then clicks OK to
-        begin reconciliation.
-
-        Parameters
-        ----------
-        account_name
-            Account to reconcile (must match an item in the combo).
-        statement_date
-            Statement end date (e.g. "03/31/2026").
-        ending_balance
-            Statement ending balance (e.g. "1,234.00").
-        service_charge, service_date, interest_earned, interest_date
-            Optional bank-charge and interest fields.
-        timeout_ms
-            Max wait (ms) for each dialog to appear.
-
-        Returns
-        -------
-        dict
-            ``{"ok": True, "account": str, "statement_date": str,
-               "ending_balance": str}``
-        """
-        import ctypes  # noqa: PLC0415
-        import ctypes.wintypes  # noqa: PLC0415
-        import time  # noqa: PLC0415
-
-        user32 = ctypes.windll.user32
-
-        from server.process_manager import get_process_manager  # noqa: PLC0415
-        pm = get_process_manager()
-        if not pm.attached:
-            raise TargetNotFoundError("Use select_window to attach first.")
-
-        WM_COMMAND = 0x0111
-        WM_GETTEXT = 0x000D
-        WM_SETTEXT = 0x000C
-        WM_GETTEXTLENGTH = 0x000E
-        CB_GETCOUNT = 0x0146
-        CB_GETLBTEXT = 0x0148
-        CB_GETLBTEXTLEN = 0x0149
-        CB_SETCURSEL = 0x014E
-        BM_CLICK = 0x00F5
-
-        root_hwnd = pm.attached.hwnd
-
-        def _read_text(h: int) -> str:
-            tlen = user32.SendMessageW(h, WM_GETTEXTLENGTH, 0, 0)
-            if tlen <= 0:
-                return ""
-            buf = ctypes.create_unicode_buffer(tlen + 1)
-            user32.SendMessageW(h, WM_GETTEXT, len(buf), buf)
-            return buf.value
-
-        def _click_qc_button(h: int) -> None:
-            """Click a QC_button via WM_LBUTTONDOWN/WM_LBUTTONUP."""
-            rc = ctypes.wintypes.RECT()
-            user32.GetClientRect(h, ctypes.byref(rc))
-            cx = (rc.right - rc.left) // 2
-            cy = (rc.bottom - rc.top) // 2
-            lp = ctypes.c_long((cy << 16) | (cx & 0xFFFF)).value
-            user32.SetFocus(h)
-            user32.SendMessageW(h, 0x0201, 1, lp)  # WM_LBUTTONDOWN
-            user32.SendMessageW(h, 0x0202, 0, lp)  # WM_LBUTTONUP
-
-        def _wait_for_dialog(title_substr: str, timeout: float) -> int | None:
-            """Poll until a top-level dialog with matching title appears."""
-            buf = ctypes.create_unicode_buffer(256)
-            deadline = time.monotonic() + timeout
-            while time.monotonic() < deadline:
-                found: list[int] = []
-                def _enum(h: int, _: int) -> bool:
-                    if user32.IsWindowVisible(h):
-                        user32.GetWindowTextW(h, buf, 256)
-                        if title_substr.lower() in buf.value.lower():
-                            found.append(h)
-                    return True
-                cb = ctypes.WINFUNCTYPE(
-                    ctypes.c_bool, ctypes.wintypes.HWND, ctypes.wintypes.LPARAM
-                )(_enum)
-                user32.EnumWindows(cb, 0)
-                if found:
-                    return found[0]
-                time.sleep(0.15)
-            return None
-
-        def _get_dialog_children(dlg_hwnd: int) -> list[tuple[int, str, str]]:
-            buf = ctypes.create_unicode_buffer(256)
-            items: list[tuple[int, str, str]] = []
-            def _ec(h: int, _: int) -> bool:
-                cls = ctypes.create_unicode_buffer(64)
-                user32.GetClassNameW(h, cls, 64)
-                user32.GetWindowTextW(h, buf, 256)
-                items.append((h, cls.value.lower(), buf.value))
-                return True
-            cb = ctypes.WINFUNCTYPE(
-                ctypes.c_bool, ctypes.wintypes.HWND, ctypes.wintypes.LPARAM
-            )(_ec)
-            user32.EnumChildWindows(dlg_hwnd, cb, 0)
-            return items
-
-        timeout_s = timeout_ms / 1000.0
-
-        # ── Step 1: send WM_COMMAND 7203 to open Choose Reconcile Account ──
-        user32.PostMessageW(root_hwnd, WM_COMMAND, 7203, 0)
-
-        # Wait for either the Choose Account dialog OR the "no items" notification.
-        # Poll for both in parallel since the "no uncleared items" dialog has
-        # title starting with "Quicken" (no fixed title for detection).
-        acct_dlg: int | None = None
-        already_done: bool = False
-        buf_tmp = ctypes.create_unicode_buffer(512)
-        deadline_step1 = time.monotonic() + timeout_s
-        while time.monotonic() < deadline_step1:
-            hits: list[int] = []
-            def _poll(h: int, _: int) -> bool:
-                if user32.IsWindowVisible(h):
-                    user32.GetWindowTextW(h, buf_tmp, 512)
-                    t_low = buf_tmp.value.lower()
-                    if "choose reconcile account" in t_low:
-                        hits.append(h)
-                    elif "no uncleared" in t_low or "nothing to reconcile" in t_low:
-                        hits.append(-(h))  # negative = already-done dialog
-                return True
-            cb_poll = ctypes.WINFUNCTYPE(
-                ctypes.c_bool, ctypes.wintypes.HWND, ctypes.wintypes.LPARAM
-            )(_poll)
-            user32.EnumWindows(cb_poll, 0)
-            for hit in hits:
-                if hit > 0:
-                    acct_dlg = hit
-                    break
-                else:
-                    # "no uncleared items" — dismiss and report
-                    already_done = True
-                    ok_dismiss = next(
-                        (h for h, c, t in _get_dialog_children(-hit)
-                         if c == "qc_button" and t == "OK"),
-                        None,
-                    )
-                    if ok_dismiss:
-                        _click_qc_button(ok_dismiss)
-                    break
-            if acct_dlg or already_done:
-                break
-            time.sleep(0.15)
-
-        if already_done:
-            return {
-                "ok": True,
-                "account": account_name,
-                "statement_date": statement_date,
-                "ending_balance": ending_balance,
-                "note": "Reconcile already active or no uncleared items.",
-            }
-
-        if acct_dlg is None:
-            raise UIAError(
-                "Choose Reconcile Account dialog did not appear.",
-                code="DIALOG_NOT_FOUND",
-            )
-
-        children = _get_dialog_children(acct_dlg)
-
-        # Find the combo and select the account.
-        combo_h = next((h for h, c, _ in children if c == "qwcombobox"), None)
-        if combo_h is None:
-            raise UIAError("Account combo not found in reconcile dialog.", code="ELEMENT_NOT_FOUND")
-
-        count = user32.SendMessageW(combo_h, CB_GETCOUNT, 0, 0)
-        target_idx: int | None = None
-        for i in range(count):
-            tlen = user32.SendMessageW(combo_h, CB_GETLBTEXTLEN, i, 0)
-            tbuf = ctypes.create_unicode_buffer(tlen + 2)
-            user32.SendMessageW(combo_h, CB_GETLBTEXT, i, tbuf)
-            if tbuf.value.strip().lower() == account_name.strip().lower():
-                target_idx = i
-                break
-
-        if target_idx is None:
-            # Cancel the dialog and raise.
-            cancel_h = next((h for h, c, t in children if c == "qc_button" and t == "Cancel"), None)
-            if cancel_h:
-                _click_qc_button(cancel_h)
-            raise UIAError(
-                f"Account '{account_name}' not found in reconcile combo.",
-                code="ACCOUNT_NOT_FOUND",
-            )
-
-        user32.SendMessageW(combo_h, CB_SETCURSEL, target_idx, 0)
-
-        # Click OK.
-        ok_h = next((h for h, c, t in children if c == "qc_button" and t == "OK"), None)
-        if ok_h is None:
-            raise UIAError("OK button not found in account selection dialog.", code="ELEMENT_NOT_FOUND")
-        _click_qc_button(ok_h)
-
-        # ── Step 2: wait for Reconcile Details dialog ──
-        det_dlg = _wait_for_dialog("Reconcile Details", timeout_s)
-        if det_dlg is None:
-            raise UIAError(
-                "Reconcile Details dialog did not appear after account selection.",
-                code="DIALOG_NOT_FOUND",
-            )
-
-        det_children = _get_dialog_children(det_dlg)
-        edit_handles = [h for h, c, _ in det_children if c == "edit"]
-        # Layout order (by z-order): date, prior balance, ending balance,
-        # service_charge, service_date, service_category,
-        # interest_earned, interest_date, interest_category.
-        # Indices 0, 2 are the ones we must fill.
-        if len(edit_handles) < 3:
-            raise UIAError(
-                f"Expected ≥3 Edit fields in Reconcile Details, got {len(edit_handles)}.",
-                code="UNEXPECTED_LAYOUT",
-            )
-
-        def _set_edit(h: int, text: str) -> None:
-            if not text:
-                return
-            user32.SetFocus(h)
-            user32.SendMessageW(h, WM_SETTEXT, 0, text)
-
-        _set_edit(edit_handles[0], statement_date)   # Ending statement date
-        _set_edit(edit_handles[2], ending_balance)   # Ending balance
-        if service_charge:
-            _set_edit(edit_handles[3], service_charge)
-        if service_date:
-            _set_edit(edit_handles[4], service_date)
-        if interest_earned:
-            _set_edit(edit_handles[6], interest_earned)
-        if interest_date:
-            _set_edit(edit_handles[7], interest_date)
-
-        # Click OK to begin reconciliation.
-        det_ok_h = next(
-            (h for h, c, t in det_children if c == "qc_button" and t == "OK"), None
-        )
-        if det_ok_h is None:
-            raise UIAError("OK button not found in Reconcile Details dialog.", code="ELEMENT_NOT_FOUND")
-        _click_qc_button(det_ok_h)
-
-        time.sleep(0.5)
-
-        return {
-            "ok": True,
-            "account": account_name,
-            "statement_date": statement_date,
-            "ending_balance": ending_balance,
         }
